code: refactoring login page
